Adding project files
Adding implementation and header files. These include  the following:
the coordinate ADT to objectify the grid cells; the coordStack ADT which
stores and manages the aforementioned coordinate objects as a stack,
which is implemented as a singly linked list; the maze header which
contains the solution for solving the mazes. See files for further
details of project. Also included are sample input and output files.